diary: error messages
