+++ controllers/diary/removeExercise.js
@@ -28,10 +28,21 @@ const removeExercise = async (req, res) => {
     throw HttpError(401, "This exercise is not found");
   }
 
+  // const result = await Diary.findByIdAndUpdate(
+  //   foundDiary._id,
+  //   {
+  //     $inc: {
+  //       timeSport: -completedExercises.time,
+  //       burnedCalories: -completedExercises.calories,
+  //     },
+  //     $pull: { doneExercises: { _id: completedExercises._id } },
+  //   },
+  //   { new: true }
+  // ).select("-createdAt -updatedAt");
   const result = await Diary.findByIdAndUpdate(
     foundDiary._id,
     {
-      $inc: { timeSport: -completedExercises.time, burnedCalories: -completedExercises.calories },
+      $inc: { burnedCalories: -completedExercises.calories },
       $pull: { doneExercises: { _id: completedExercises._id } },
     },
     { new: true }
@@ -39,5 +50,8 @@ const removeExercise = async (req, res) => {
 
   res.json(result);
 };
+
+//   res.json(result);
+// };
 
 module.exports = removeExercise;
