--- helpers/errorAmountMessages.js
@@ -1,7 +1,0 @@
-const errorAmountMessages = {
-  "number.base": "The amount must be a number.",
-  "any.required": "The amount field is required.",
-  "number.min": "The amount must be greater than 0.",
-};
-
-module.exports = errorAmountMessages;
