--- helpers/errorDateMessages.js
@@ -1,9 +1,0 @@
-const errorDateMessages = {
-  "string.base": "The date must be a string.",
-  "any.required": "The date field is required.",
-  "string.empty": "The date must not be empty",
-  "string.pattern.base": 'The date must be in the format "DD-MM-YYYY"',
-  "date.inPast": "Date must not be in the future",
-};
-
-module.exports = errorDateMessages;
