--- helpers/errorExerciseMessages.js
@@ -1,7 +1,0 @@
-const errorЕxerciseMessages = {
-  "string.base": "The exercise must be a string.",
-  "any.required": "The exercise field is required.",
-  "string.empty": "The exercise  must not be empty.",
-};
-
-module.exports = errorЕxerciseMessages;
